apps/worker/requirements.txt
boto3==1.35.24
python-dotenv==1.0.1
pytest==8.3.2
ruff==0.6.8
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
pillow==10.4.0
pymupdf==1.24.10
